backend/requirements.txt
# FlexVerify Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File Uploads
python-multipart==0.0.9

# Face Recognition (dlib-based, no GPU required)
face-recognition==1.3.0

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.24.3
pillow==10.2.0
scikit-image==0.22.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
